lambda/requirements.txt
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.14.0
boto3>=1.26.0
botocore>=1.29.0 